+++ kubernetes/tests/test.sh
@@ -34,17 +34,12 @@ cp -rf ../app ~/app
 sleep 50
 kubectl get -n lnmp all
 curl -k https://laravel2.t.khs1994.com
-ip addr
-test "${LNMP_K8S_SINGLE_INSTALL_OPTIONS}" = "--containerd" && (kubectl apply -f deployment/runtimeClass/runtimeClass.yaml && kubectl apply -f deployment/runtimeClass/pod.yaml) || true
-test "${LNMP_K8S_SINGLE_INSTALL_OPTIONS}" = "--crio" && (kubectl apply -f deployment/runtimeClass/runtimeClass.yaml && kubectl apply -f deployment/runtimeClass/pod.yaml) || true
-test -z "${LNMP_K8S_SINGLE_INSTALL_OPTIONS}" && docker run -it --rm --runtime=runsc alpine:3.10 uname -a || true
-test -z "${LNMP_K8S_SINGLE_INSTALL_OPTIONS}" && docker run -it --rm alpine:3.10 uname -a || true
+test "${LNMP_K8S_LOCAL_INSTALL_OPTIONS}" = "--containerd" && (kubectl apply -f deployment/runtimeClass/runtimeClass.yaml && kubectl apply -f deployment/runtimeClass/pod.yaml) || true
+test "${LNMP_K8S_LOCAL_INSTALL_OPTIONS}" = "--crio" && (kubectl apply -f deployment/runtimeClass/runtimeClass.yaml && kubectl apply -f deployment/runtimeClass/pod.yaml) || true
+test -z "${LNMP_K8S_LOCAL_INSTALL_OPTIONS}" && docker run -it --rm --runtime=runsc alpine:3.10 uname -a || true
+test -z "${LNMP_K8S_LOCAL_INSTALL_OPTIONS}" && docker run -it --rm alpine:3.10 uname -a || true
 sleep 10
 kubectl get all
 kubectl get pod
 POD_NAME=`kubectl get pod | awk '{print $1}' | tail -1` || true
 kubectl exec ${POD_NAME} -- uname -a || true
-cd cli
-sh generate.sh || true
-cd ..
-git diff