+++ kubernetes/CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+* CRI 默认为 `containerd`, 还有 `crio` `docker` 可供选择
+
 * Kubernetes `1.16.1`
 
 * Helm 3