+++ kubernetes/CHANGELOG.md
@@ -2,6 +2,6 @@
 
 * CRI 默认为 `containerd`, 还有 `crio` `docker` 可供选择
 
-* Kubernetes `1.16.1`
+* Kubernetes `1.17.0`
 
 * Helm 3