+++ kubernetes/CHANGELOG.md
@@ -5,3 +5,7 @@
 * Kubernetes `1.18.0`
 
 * Helm 3
+
+## v1.19.0-alpha.2
+
+* `kubescheduler.config.k8s.io/v1alpha1` 替换为 `kubescheduler.config.k8s.io/v1alpha2`